fix: page size long 형태에서 int 로 변경. 참여 이력 페이징 조회 서비스 로직 테스트 코드 작성

+++ src/main/kotlin/com/task/kakaopayadvertisementserver/controller/admin/AdvertisementController.kt
@@ -37,7 +37,7 @@ class AdvertisementController(
     @GetMapping
     fun findPagedAdvertisements(
         @RequestParam(defaultValue = "$DEFAULT_PAGE") @Min(0) page: Int,
-        @RequestParam(defaultValue = "$MAX_ADVERTISEMENT_PAGE_SIZE") @Min(1) @Max(MAX_ADVERTISEMENT_PAGE_SIZE) size: Int,
+        @RequestParam(defaultValue = "$MAX_ADVERTISEMENT_PAGE_SIZE") @Min(1) @Max(MAX_ADVERTISEMENT_PAGE_SIZE.toLong()) size: Int,
     ) {
         advertisementService.findPagedAdvertisement(page, size, LocalDateTime.now())
     }
